feature: 1. Connect to a database in config 2. add seperate model file to introduceall the new files 3. add __repr__ to the model plus adding the show object
