+++ projects/01_fyyur/starter_code/config.py
@@ -10,4 +10,7 @@ DEBUG = True
 
 
 # TODO IMPLEMENT DATABASE URL
-SQLALCHEMY_DATABASE_URI = '<Put your local database url>'
+SQLALCHEMY_DATABASE_URI = 'postgresql://postgres@localhost:5432/fyyur'
+
+# remove console warning
+SQLALCHEMY_TRACK_MODIFICATIONS = False
