+++ projects/01_fyyur/starter_code/models/model.py
@@ -1,0 +1,78 @@
+#----------------------------------------------------------------------------#
+# Imports.
+#----------------------------------------------------------------------------#
+import json
+from flask import Flask
+from flask_migrate import Migrate
+from flask_sqlalchemy import SQLAlchemy
+from datetime import datetime
+
+
+#----------------------------------------------------------------------------#
+# Models.
+#----------------------------------------------------------------------------#
+
+app = Flask(__name__)
+app.config.from_object('config')
+db = SQLAlchemy(app)
+
+# TODO: connect to a local postgresql database
+migrate = Migrate(app, db)
+
+#----------------------------------------------------------------------------#
+# Models.
+#----------------------------------------------------------------------------#
+
+class Venue(db.Model):
+    __tablename__ = 'Venue'
+
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String)
+    city = db.Column(db.String(120))
+    state = db.Column(db.String(120))
+    address = db.Column(db.String(120))
+    phone = db.Column(db.String(120))
+    image_link = db.Column(db.String(500))
+    facebook_link = db.Column(db.String(120))
+
+    # TODO: implement any missing fields, as a database migration using Flask-Migrate
+    # TODO: add functions needed as insert, update, delete, description
+    # TODO: Add DB relationship
+    def __repr__(self):
+      return f'<Artist {self.id} {self.name} >'
+
+class Artist(db.Model):
+    __tablename__ = 'Artist'
+
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String)
+    city = db.Column(db.String(120))
+    state = db.Column(db.String(120))
+    phone = db.Column(db.String(120))
+    genres = db.Column(db.String(120))
+    image_link = db.Column(db.String(500))
+    facebook_link = db.Column(db.String(120))
+
+    # TODO: implement any missing fields, as a database migration using Flask-Migrate
+    # add the db relationship
+    # TODO: implement any missing fields, as a database migration using Flask-Migrate
+    # TODO: add functions needed as insert, update, delete, description
+    # TODO: Add DB relationship
+    def __repr__(self):
+      return f'<Venue {self.id} {self.name} >'
+
+#TODO Implement Show and Artist models, and complete all model relationships and properties, as a database migration.
+class Show(db.Model):
+    __tablename__ = 'Show'
+
+    id = db.Column(db.Integer, primary_key=True)
+    artist_id = db.Column(db.Integer, db.ForeignKey('Artist.id'),nullable=False)
+    venue_id = db.Column(db.Integer, db.ForeignKey('Venue.id'), nullable=False)
+    start_time = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
+    # TODO: implement any missing fields, as a database migration using Flask-Migrate
+    # add the db relationship
+    # TODO: implement any missing fields, as a database migration using Flask-Migrate
+    # TODO: add functions needed as insert, update, delete, description
+    # TODO: Add DB relationship
+    def __repr__(self):
+      return f'<Show {self.id}, Artist {self.artist_id}, Venue {self.venue_id} and Start Time {self.start_time}>'
